requirements.txt
absl-py==2.0.0
flax==0.10.0
gym==0.23.1
matplotlib==3.8.2
ml_collections==0.1.1
tensorflow-probability==0.23.0
distrax==0.1.5
tqdm==4.66.1
gdown==5.2.0
scipy==1.11.4
wandb==0.16.1
cython==0.29.36
patchelf==0.17.2.1
d4rl @ git+https://github.com/wilcoxsonm21/d4rl@master # also installs mujoco-py, dm_control, and other d4rl dependencies
jax[cuda]>=0.4.16
dm_control<=1.0.20
mujoco<=3.1.6
imageio==2.36.0
moviepy==1.0.3
